Added Edit and Delete Contact functionalities, triggered by clicking on a contact

+++ app/src/main/java/edu/msu/cse476/msucompanion/Contact.java
@@ -20,10 +20,10 @@ public class Contact {
     }
 
     // Getters and Setters
-    public long getId() { return id; }
+    public int getId() { return id; }
     public void setId(int id) { this.id = id; }
 
-    public long getUserId() { return userId; }
+    public int getUserId() { return userId; }
     public void setUserId(int userId) { this.userId = userId; }
 
     public String getName() { return name; }
